Refactor project to run on LibGDX game library
This commit marks a major code overhaul where the heavy lifting is
managed by the libGDX game engine library. In this version, upon
running, five attackers walk across the screen and are "killed" when
tapped. The rectangles wrapping the characters are too large currently,
so overlap occurs.

Attackers are enemies in the game that have their own sprite object and
rectangle. They do not automatically have animation, so in order to
create an attacker successfully you MUST construct a new attacker, add
at least one animation (mStandardAttackerWalk) and then invoke revive()
on that attacker to actually see it.